udn, tests: allow addNetwork to be reused for primary UDNs
The add network function can now add network controllers whose default
cluster network controller does *not* have the gateway.

Also, it is now possible to define the default route in a secondary
network controller.

Expected routes and role can be defined for each of the controllers
via the shared `testPod` struct.
