+++ go-controller/pkg/ovn/multihoming_test.go
@@ -1,0 +1,43 @@
+package ovn
+
+import "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
+
+func (p testPod) addNetwork(
+	netName, nadName, nodeSubnet, nodeMgtIP, nodeGWIP, podIP, podMAC, role string,
+	tunnelID int,
+	routes []util.PodRoute,
+) {
+	podInfo, ok := p.secondaryPodInfos[netName]
+	if !ok {
+		podInfo = &secondaryPodInfo{
+			nodeSubnet:  nodeSubnet,
+			nodeMgtIP:   nodeMgtIP,
+			nodeGWIP:    nodeGWIP,
+			role:        role,
+			routes:      routes,
+			allportInfo: map[string]portInfo{},
+		}
+		p.secondaryPodInfos[netName] = podInfo
+	}
+	portName := util.GetSecondaryNetworkLogicalPortName(p.namespace, p.podName, nadName)
+	podInfo.allportInfo[nadName] = portInfo{
+		portUUID: portName + "-UUID",
+		podIP:    podIP,
+		podMAC:   podMAC,
+		portName: portName,
+		tunnelID: tunnelID,
+	}
+}
+
+func (p testPod) getNetworkPortInfo(netName, nadName string) *portInfo {
+	podInfo, ok := p.secondaryPodInfos[netName]
+	if !ok {
+		return nil
+	}
+	info, ok := podInfo.allportInfo[nadName]
+	if !ok {
+		return nil
+	}
+
+	return &info
+}
