+++ go-controller/pkg/ovn/pods_test.go
@@ -201,6 +201,7 @@ type secondaryPodInfo struct {
 	nodeSubnet  string
 	nodeMgtIP   string
 	nodeGWIP    string
+	role        string
 	routes      []util.PodRoute
 	allportInfo map[string]portInfo
 }
@@ -342,6 +343,13 @@ func (p testPod) getAnnotationsJson() string {
 	}
 
 	for _, portInfos := range p.secondaryPodInfos {
+		var secondaryIfaceRoutes []podRoute
+		for _, route := range portInfos.routes {
+			secondaryIfaceRoutes = append(
+				secondaryIfaceRoutes,
+				podRoute{Dest: route.Dest.String(), NextHop: route.NextHop.String()},
+			)
+		}
 		for nad, portInfo := range portInfos.allportInfo {
 			ipPrefix := 24
 			if ovntest.MustParseIP(p.podIP).To4() == nil {
@@ -353,6 +361,12 @@ func (p testPod) getAnnotationsJson() string {
 				IP:       ip,
 				IPs:      []string{ip},
 				TunnelID: portInfo.tunnelID,
+				Role:     portInfos.role,
+				Routes:   secondaryIfaceRoutes,
+			}
+			if portInfos.nodeGWIP != "" {
+				podAnnotation.Gateway = portInfos.nodeGWIP
+				podAnnotation.Gateways = []string{portInfos.nodeGWIP}
 			}
 			podAnnotations[nad] = podAnnotation
 		}
